upnp: Gracefully handle multi-homed devices
If a device is reachable via different proxies (since it's available on
different network interfaces) we "ref" the device by adding the new proxy.
Device interaction is done using the first proxy we got for the device.

If a proxy disappears, it will be removed from the device's proxy list; if
it's the controlling proxy we use, we pick the next one from the list.